update comments and file naming
